requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect==1.3.0
distro==1.9.0
ecdsa==0.19.1
fake-useragent==2.2.0
fastapi==0.115.12
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-texttospeech==2.27.0
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.9.0
joblib==1.5.0
jpype1==1.5.2
konlpy==0.6.0
krwordrank==1.0.3
loguru==0.7.2
lxml==5.4.0
lxml_html_clean==0.4.2
ngrok==1.3.0
numpy==2.2.5
openai==1.76.0
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
passlib==1.7.4
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
requests-html==0.10.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
starlette==0.46.2
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.34.2
w3lib==2.3.1
websocket-client==1.8.0
websockets==10.4
win32_setctime==1.2.0
wsproto==1.2.0
zipp==3.22.0
